miscellaneous updates to MTA:GREEN; improved the OOP environment; fixed Lua networking (CLuaArguments); medi-STABLE (debug this build please, especially it's Lua environment!); if you experience the deathmatch error, recompile everything and try running the build again!

+++ Shared/sdk/CQuat.h
@@ -14,14 +14,13 @@ class CQuat;
 #ifndef __CQUAT_H
 #define __CQUAT_H
 
-#include <math.h>
 #include <float.h>
 #include "SharedUtil.h"
 
 #define EPSILON         0.001f
 
 #ifndef PI
-#define PI (float)3.14159265358979323846264338327950
+#define PI 3.14159265358979323846264338327950f
 #endif
 
 class CQuat
